chore: move current epoch metadata query into useCurrentEpochMetadataQuery
